+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Bugfix: preview buttons weren't working from editors in the Issues page
 * Links in rich text now open in new tabs
+* Automatically update the Issues page filter when the project changes
 
 # 3.1.0